requirements.txt
annotated-types==0.6.0
anyio==4.2.0
async-timeout==4.0.3
beautifulsoup4==4.12.3
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
exceptiongroup==1.2.0
fastapi==0.108.0
fastapi-cache2==0.2.1
filelock==3.13.3
gdown==5.1.0
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
idna==3.6
iniconfig==2.0.0
joblib==1.3.2
lightfm==1.17
numpy==1.26.3
packaging==24.0
pandas==2.1.4
pendulum==3.0.0
pluggy==1.4.0
pydantic==2.5.3
pydantic_core==2.14.6
PySocks==1.7.1
pytest==8.1.1
pytest-asyncio==0.23.6
python-dateutil==2.8.2
python-dotenv==1.0.0
python-multipart==0.0.6
pytz==2023.3.post1
PyYAML==6.0.1
redis==4.6.0
requests==2.31.0
scikit-learn==1.3.2
scipy==1.11.4
six==1.16.0
sniffio==1.3.0
soupsieve==2.5
starlette==0.37.2
threadpoolctl==3.2.0
time-machine==2.14.0
tomli==2.0.1
tqdm==4.66.1
typing_extensions==4.9.0
tzdata==2023.4
urllib3==2.1.0
uvicorn==0.25.0
uvloop==0.19.0
watchfiles==0.21.0
websockets==12.0
